% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zuxlsx-package.R
\docType{package}
\name{zuxlsx-package}
\alias{zuxlsx}
\alias{zuxlsx-package}
\title{zuxlsx: Read 'xlsx' Workbooks Without System Dependencies}
\description{
Reads spreadsheet data from 'xlsx' workbooks using a bundled copy of the 'xlsxio' reader (\url{https://github.com/brechtsanders/xlsxio}), so that no system XML or ZIP library is required. The 'Expat' parser, the 'miniz' ZIP reader and the 'TF-PSA-Crypto' cryptography used to open password-protected workbooks are linked statically at install time from the 'zuxml', 'zukomp' and 'zucrypt' packages, which means there is no run-time dependency on any of them. Worksheets are read as a stream, so a workbook does not have to be held in memory in full, unless it is password-protected and has to be decrypted first. Writing workbooks is out of scope.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pedrobtz/zuxlsx}
  \item \url{https://pedrobtz.github.io/zuxlsx/}
  \item Report bugs at \url{https://github.com/pedrobtz/zuxlsx/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Brecht Sanders (xlsxio, bundled in src/vendor/xlsxio) [copyright holder]
}

}
\keyword{internal}
