% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets.R
\name{xlsx_sheets}
\alias{xlsx_sheets}
\title{List the worksheets in an xlsx workbook}
\usage{
xlsx_sheets(path)
}
\arguments{
\item{path}{Path to an \code{.xlsx} file.}
}
\value{
A character vector of worksheet names, in workbook order. A workbook
always has at least one.
}
\description{
List the worksheets in an xlsx workbook
}
\examples{
path <- system.file("extdata", "two-sheets.xlsx", package = "zuxlsx")
xlsx_sheets(path)
}
\seealso{
\link{zuxlsx-conditions} for the errors this can raise.
}
