% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zuxlsx-package.R
\docType{package}
\name{zuxlsx-package}
\alias{zuxlsx}
\alias{zuxlsx-package}
\title{zuxlsx: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/zuxlsx/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
