% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zuxlsx-package.R
\docType{package}
\name{zuxlsx-package}
\alias{zuxlsx}
\alias{zuxlsx-package}
\title{zuxlsx: Read 'xlsx' Workbooks Without System Dependencies}
\description{
Reads spreadsheet data from 'xlsx' workbooks using a bundled copy of the 'xlsxio' reader (\url{https://github.com/brechtsanders/xlsxio}), so that no system XML or ZIP library is required. The 'Expat' parser and the 'miniz' ZIP reader are linked statically at install time from the 'zuxml' and 'zukomp' packages, which means there is no run-time dependency on either of them. Worksheets are read as a stream, so a workbook does not have to be held in memory in full. Writing workbooks is out of scope.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pedrobtz/zuxlsx}
  \item \url{https://pedrobtz.github.io/zuxlsx/}
  \item Report bugs at \url{https://github.com/pedrobtz/zuxlsx/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Brecht Sanders (xlsxio, bundled in src/vendor/xlsxio) [copyright holder]
}

}
\keyword{internal}
